Moved code into new file, updated interface
