+++ components/PostItem.tsx
@@ -1,11 +1,44 @@
+import API, { graphqlOperation } from "@aws-amplify/api";
 import { APIListOperations } from "@components/APIList";
-import React, { useState, useEffect } from "react";
-import { StyleSheet, Text, View } from "react-native";
+import { deletePost, updatePost } from "@graphql/mutations";
+import React, { useRef, useState } from "react";
+import { Alert, Dimensions, FlatList, StyleSheet, Text, TextInput, TouchableOpacity, View } from "react-native";
 import { Post } from "src/models";
 import IconButton from "./common/IconButton";
+import Modal, { ModalRefType } from "./common/Modal";
+import CommentsModal from "./postComponents/CommentsModal";
+import LikesModal from "./postComponents/LikesModal";
+import PostHeader from "./postComponents/PostHeader";
+import PostImage from "./PostImage";
 import { ProfileImageAndName } from "./ProfileImageAndName";
-import { Divider } from "react-native-elements";
-import generateColor from "@hooks/generateRandomColor";
+import TextWithTaggedUsers from "./TextWithTaggedUsers";
+
+
+const updatePostAWS = async (createdAt: string, editedText: string) => {
+  try {
+    await API.graphql(
+      graphqlOperation(updatePost, {
+        input: { createdAt: createdAt, description: editedText },
+      })
+    );
+    console.log("success in updating a post");
+  } catch (err) {
+    console.warn("error in updating post: ", err);
+  }
+};
+
+const deletePostAWS = async (createdAt: string) => {
+  try {
+    await API.graphql(
+      graphqlOperation(deletePost, {
+        input: { createdAt: createdAt, userId: globalThis.myId },
+      })
+    );
+    console.log("success in deleting a post");
+  } catch {
+    console.log("error in deleting post: ");
+  }
+};
 
 interface Props {
   item: Post & {taggedUsers?: string[]; likedByYou?: boolean},
@@ -16,9 +49,6 @@ interface Props {
   isVisible?: boolean,
   shouldSubscribe?: boolean,
   operations: APIListOperations<Post>,
-  startTime?: Date, 
-  maxOccupancy?: number,
-  hasInvitations: boolean
 }
 
 const PostItem = ({
@@ -35,185 +65,202 @@ const PostItem = ({
   //replies,
   //index,
   operations,
-  startTime,
-  maxOccupancy,
-  hasInvitations,
 } : Props) => {
-  const [requested, setRequested] = useState(false); // If user has requested to join
-  const [numInvitations, setNumInvitations] = useState(0) // Number of requested invitations
-  const [isHours, setIsHours] = useState(true); // If time limit has >= 1 hour left
-  const [timeUntil, setTimeUntil] = useState(0);
-  const [currentCapacity, setCurrentCapacity] = useState(5);
-  const [color, setColor] = useState('black');
-  const NUM_OF_LINES = 5;
-  const CAPACITY_PERCENTAGE = 0.75;
-
-  const setTime = () => {
-    if (startTime) {
-      const date = new Date()
-      const diffTime = Math.abs(startTime.getTime() - date.getTime());
-      const diffMin = Math.ceil(diffTime / (1000 * 60));
-      const diffHour = Math.ceil(diffTime / (1000 * 60 * 60));
-
-      if (diffMin < 60) {
-        setIsHours(false);
-        setTimeUntil(diffMin);
-
-      } else {
-        setIsHours(true);
-        setTimeUntil(diffHour);
-      }
-    } 
-  }
-
-  useEffect(() => {
-    setColor(generateColor);
-  }, []);
-
-  useEffect(() => {
-    setTime();
-  }, [startTime]);
-
-  const handleRequestToJoin = () => {
-    if (requested) {
-      setRequested(false);
-      setNumInvitations(numInvitations-1);
-    } else {
-      setRequested(true);
-      setNumInvitations(numInvitations+1);
-    }
-  };
-
+  const {removeItem, replaceItem} = operations;
+  const likesModalRef = useRef<ModalRefType>(null);
+  const repliesModalRef = useRef<ModalRefType>(null);
+  const [isEditing, setIsEditing] = useState(false);
+  const [editedText, setEditedText] = useState("");
+  
   return (
     <View style={styles.secondaryContainerStyle}>
       <View
         style={[styles.spaceAround, replyButtonHandler ? {} : styles.nestedReply]}
       >
-      {/* Header (name, profile pic, event dot, distance) */}
+        <PostHeader
+          item={item}
+          writtenByYou={writtenByYou}
+          toggleEditing={() => setIsEditing(!isEditing)}
+          repliesPressed={() => {
+            repliesModalRef.current?.showModal();
+          }}
+          reportPost={reportPost}
+          shouldSubscribe={shouldSubscribe}
+        />
+
+        <PostImage
+          style={{
+            resizeMode: "cover",
+            width: Dimensions.get("window").width,
+            height: Dimensions.get("window").width,
+            alignSelf: "center",
+            marginBottom: 15,
+          }}
+          filename={item.imageURL}
+          isVisible={
+            isVisible
+            // && !areRepliesVisible
+          }
+        />
+
         <View
-          style={[styles.flexRow, {paddingBottom: '2%'}]}
+          style={{
+            flexDirection: "row",
+            justifyContent: "space-between",
+            minHeight: writtenByYou ? 70 : 35,
+          }}
         >
-          <ProfileImageAndName
-            textStyle={{
-              fontWeight: writtenByYou ? "bold" : "normal",
-            }}
-            style={styles.profile}
-            userId={item.userId}
-          />
-          <IconButton
-              style={styles.eventDot}
-              iconName={"lens"}
-              size={15}
-              color={color}
-              onPress={() => null}
-            />
-          <Text style={styles.distance}>0 mi</Text>
-        </View>
+          {isEditing ? (
+            <TextInput
+              style={[styles.check, { borderColor: "orange" }]}
+              onChangeText={setEditedText}
+              autoFocus={true}
+            >
+              {item.description}
+            </TextInput>
+          ) : (
+            <TextWithTaggedUsers textInput={item.description} taggedUsers={item.taggedUsers} urlPreview={item.urlPreview}/>
+          )}
 
-        <Divider style={styles.divider}/>
-
-        {/* Description */}
-        <View style={styles.description}>
-          <Text numberOfLines={NUM_OF_LINES}
-            style={{
-              paddingHorizontal: '3%'
-            }}
-            accessibilityLabel={'description'}
+          <View
+            style={{ flexDirection: "column", position: "absolute", right: 15 }}
           >
-            {item.description}
-          </Text>
-        </View>
-
-        {/* Bottom Left Icons (time until event, max occupancy) */}
-        <View style={[styles.flexRow, {paddingBottom: '1%'}]}>
-          <View style={[styles.flexRow, {paddingLeft: '2%'}]}>
-            {startTime != null ?
-              <View style={{flexDirection: 'row'}} accessibilityLabel={'time until'}>
-                <IconButton
-                  iconName={"query-builder"}
-                  size={22}
-                  color={isHours ? 'grey' : 'red'}
-                  onPress={() => null}
-                  accessibilityLabel={'time icon'}
-                />
-                <Text style={[
-                    styles.numHours,
-                    {color: isHours ? 'grey' : 'red'}
-                  ]}
-                >{timeUntil}{isHours ? 'hrs' : 'min'}
-                </Text>
-              </View> : null
-            }
-            {startTime && maxOccupancy != null ?
+            {!writtenByYou ? (
               <IconButton
-                style={styles.paddingDot}
-                iconName={"lens"}
-                size={7}
-                color={"grey"}
-                onPress={() => null}
-              /> : null
-            }
-            {maxOccupancy ?
-              <View style={styles.maxLimit} accessibilityLabel={'max occupancy'}>
-                <IconButton 
-                  iconName={"person-outline"}
-                  size={22}
-                  color={(currentCapacity >= Math.floor(maxOccupancy*CAPACITY_PERCENTAGE))
-                    ? 'red' : 'grey'}
-                  onPress={() => null}
-                  accessibilityLabel={'occupancy icon'}
-                />
-                <Text style={[
-                    {textAlignVertical:'center'},
-                    {color: (currentCapacity >= Math.floor(maxOccupancy*CAPACITY_PERCENTAGE))
-                      ? 'red' : 'grey'
-                    }
-                  ]}>{currentCapacity}/{maxOccupancy}</Text>
-              </View> : null
-            }
-          </View>
+                iconName={"report"}
+                size={20}
+                color={"gray"}
+                onPress={() => {removeItem(), reportPost(item.createdAt, item.userId)}}
+              />
+            ) : null}
 
-          {/* Bottom Right Icons (invitations, comments, more tab) */}
-          <View style={styles.iconsBottomRight}>
-            {hasInvitations ?
-              <View style={styles.iconsBottomRight} accessibilityLabel={'request invitations'}>
-                <Text 
-                  style={[
-                    styles.numbersBottomRight,
-                    {color: requested ? color : "black"}
-                  ]}
-                  accessibilityLabel={'invitations requested'}
-                >{numInvitations > 0 ? numInvitations : null}</Text>
-                <IconButton
-                  style={{paddingLeft: '6%'}}
-                  iconName={"person-add"}
-                  size={22}
-                  color={requested ? color : "black"}
-                  onPress={handleRequestToJoin}
-                  accessibilityLabel={'invitation icon'}
-                />
-              </View> : null
-            }
-            <Text style={styles.numbersBottomRight}>0</Text>
-            <IconButton
-              style={{paddingLeft: '3%'}}
-              iconName={"messenger"}
-              size={18}
-              color={"black"}
-              onPress={() => null}
-              accessibilityLabel={'comments icon'}
-            />
-            <IconButton
-              style={{paddingLeft: '3%'}}
-              iconName={"more-vert"}
-              size={24}
-              color={"black"}
-              onPress={() => null}
-              accessibilityLabel={'more icon'}
-            />
+            {writtenByYou ? (
+              <IconButton
+                style={{ marginBottom: 10 }}
+                iconName={"delete-forever"}
+                size={20}
+                color={"gray"}
+                onPress={() => {removeItem(), deletePostAWS(item.createdAt);}}
+              />
+            ) : null}
+
+            {writtenByYou ? (
+              <IconButton
+                style={{ marginBottom: 15 }}
+                iconName={"edit"}
+                size={20}
+                color={"gray"}
+                onPress={() => setIsEditing(!isEditing)}
+              />
+            ) : null}
           </View>
         </View>
+
+        {/* {item.loading ?? (
+          <TouchableOpacity
+            style={{
+              flexDirection: "row",
+              alignItems: "center",
+              justifyContent: "space-between",
+              marginTop: 8,
+              marginBottom: 16,
+            }}
+            onPress={() => {
+              likesModalRef.current?.showModal();
+            }}
+          >
+            <APIList
+              style={{ margin: 0, padding: 0 }}
+              horizontal={true}
+              queryOperation={likesByPost}
+              queryOperationName={"likesByPost"}
+              filter={{
+                postId: item.createdAt + "#" + item.userId,
+                sortDirection: "DESC",
+              }}
+              initialAmount={1}
+              additionalAmount={0}
+              renderItem={({ item }) => (
+                <View style={{ flexDirection: "row", alignItems: "center" }}>
+                  <ProfileImageAndName
+                    style={{
+                      alignContent: "flex-start",
+                      alignItems: "center",
+                      alignSelf: "flex-end",
+                      justifyContent: "flex-start",
+                      flexDirection: "row",
+                      marginLeft: 15,
+                      marginRight: 5,
+                    }}
+                    imageSize={20}
+                    userId={item.userId}
+                    onPress={likesModalRef.current?.showModal}
+                  />
+                  <Text>
+                    {likes > 1 ? "and " + (likes - 1) + " others" : ""} liked
+                    this post
+                  </Text>
+                </View>
+              )}
+              keyExtractor={(item: Like) => item.userId}
+            />
+          </TouchableOpacity>
+        )} */}
+
+        <FlatList
+          data={item.taggedUsers}
+          keyExtractor={(item) => item}
+          renderItem={({ item }) => (
+            <ProfileImageAndName
+              style={{
+                alignContent: "flex-start",
+                alignItems: "center",
+                alignSelf: "flex-start",
+                justifyContent: "flex-start",
+                flexDirection: "row",
+                marginLeft: 15,
+                marginRight: 5,
+              }}
+              imageSize={20}
+              userId={item}
+            />
+          )}
+        />
+
+        <Modal ref={likesModalRef}>
+          <LikesModal item={item} />
+        </Modal>
+        <Modal ref={repliesModalRef}>
+          <CommentsModal item={item} operations={operations}/>
+        </Modal>
       </View>
+
+      {isEditing ? (
+        editedText === "" ? (
+          <TouchableOpacity
+            style={styles.unselectedButtonStyle}
+            onPress={() => {
+              Alert.alert("Post is empty!");
+            }}
+          >
+            <Text style={[styles.buttonTextStyle, { color: "gray" }]}>
+              {"Edit Post"}
+            </Text>
+          </TouchableOpacity>
+        ) : (
+          <TouchableOpacity
+            style={styles.buttonStyle}
+            onPress={() => {
+              replaceItem({description: editedText}),
+              updatePostAWS(item.createdAt, editedText),
+              setEditedText(""),
+              setIsEditing(false);
+            }}
+          >
+            <Text style={styles.buttonTextStyle}>{"Edit Post"}</Text>
+          </TouchableOpacity>
+        )
+      ) : null}
     </View>
   );
 };
@@ -222,66 +269,44 @@ export default React.memo(PostItem);
 
 const styles = StyleSheet.create({
   secondaryContainerStyle: {
-    backgroundColor: "#f7f7f7"
+    backgroundColor: "#a9efe0",
   },
   spaceAround: {
     paddingLeft: 0,
     paddingTop: 0,
     paddingRight: 0,
     paddingBottom: 0,
-    flex: 1,
-    flexDirection: 'column'
   },
-  flexRow: {
-    flex: 1,
-    flexDirection: 'row'
+  check: {
+    padding: 25,
+    marginTop: 16,
+    borderColor: "#bbb",
+    borderWidth: 2,
+    borderStyle: "solid",
   },
-  distance: {
-    flex: 1,
-    flexDirection: 'row',
-    alignSelf: 'center',
-    textAlign: 'right',
-    paddingRight: '4%',
-    paddingTop: '2%'
+  unselectedButtonStyle: {
+    borderWidth: 2,
+    borderColor: "gray",
+    alignSelf: "center",
+    backgroundColor: "transparent",
+    padding: 9,
+    borderRadius: 5,
+    marginHorizontal: 10,
   },
-  divider: {
-    width: '94%',
-    height: 1,
-    alignSelf: 'center'
+  buttonTextStyle: {
+    color: "white",
+    alignSelf: "center",
+    fontWeight: "bold",
+    fontSize: 20,
+    marginBottom: 2,
+    marginHorizontal: 6,
   },
-  description: {
-    paddingBottom: '3%',
-    paddingTop: '2%'
-  },
-  iconsBottomRight: {
-    flex: 1,
-    flexDirection: "row",
-    justifyContent: "flex-end",
-  },
-  numbersBottomRight: {
-    paddingRight: 3,
-    textAlignVertical:'center',
-  },
-  paddingDot: {
-    paddingTop: '2%',
-    paddingRight: '3%',
-    paddingLeft: '1.5%'
-  },
-  eventDot: {
-    paddingRight: '2%',
-    paddingTop: '2%'
-  },
-  profile: {
-    flexDirection: "row",
-    paddingLeft: "3%",
-    paddingTop: '2%'
-  },
-  maxLimit: {
-    flexDirection: 'row',
-  },
-  numHours: {
-    textAlignVertical:'center',
-    paddingLeft: '1%'
+  buttonStyle: {
+    alignSelf: "center",
+    backgroundColor: "blue",
+    padding: 10,
+    borderRadius: 5,
+    marginHorizontal: 6,
   },
   nestedReply: {
     marginBottom: 20,
